Implement AuthSession command
Support for authenticating from the host to the YubiHSM2.

This command corresponds to the EXTERNAL_AUTHENTICATE command within
the SCP03 protocol.

It first computes a host cryptogram, then uses the S-MAC key to compute
a MAC of the entire protocol message, a.k.a. the Command-MAC (or C-MAC).

The YubiHSM2 (or MockHSM) then checks both the host cryptogram and C-MAC
for authenticity. If they verify, the client/host is authenticated,
and the protocol moves into the "AUTHENTICATED" security level.