Command/Response traits w\ DeleteObject, GenAsymmetricKey, GetObjectInfo
This adds `Command` and `Response` traits which are used to provide a
framework for extending the library with commands.

Commands and Responses are now split out as individual structs which, in
theory, it should be possible to consume with a procedural macro for
generating all serialization logic, possibly using serde.

This PR implements 3 new YubiHSM2 commands, developed in tandem with the
command framework at the time it was being implemented:

- DeleteObject: delete an object stored in the HSM (e.g. a key)
- GenAsymmetricKey: randomly generate an asymmetric key
- GetObjectInfo: obtain information about an object