Have sessions own connectors
There's no real use for Connectors after we've established a session,
so move the connector into the Session rather than borrowing it.

This eliminates the need for a lifetime on Session.

This PR additionally has Session potentially cache the static identity
keys for reconnecting sessions, since that's the one thing we might
presumably do with a connector object that's sitting around.