Make MockHSM (and therefore all Connectors) Send-safe
The only reason `Connector` previously borrowed self mutably was for the
`MockHSM` as the `reqwest`-based `Connector` implementation is thread
safe to begin with.

This wraps the core `MockHSM` state in an `Arc<Mutex<State>>` which lets
us explicitly bound the `Connector` trait on `Send`.

Additionally this commit refactors `MockHSM` into separate files as it
has grown quite large.