Integrated HttpConnector
reqwest has a pretty onerous set of dependencies. After this PR, the
number of crates in Cargo.lock is reduced by a whopping 95!

But beyond that, one thing we'd potentially like to do in the future is
use this library from within Intel SGX. Presently SGX highly restricts
the dependencies that we can choose from, and things thare are out for
the forseeable future are tokio/mio (no epoll support) and any TLS
library besides rustls.