AuthKey type (and MockHSM support for `put_auth_key`)
Factors an `AuthKey` type out of the old "StaticKeys" type used by the
`securechannel` module. This was an `AuthKey` all along, it just had a
weird name because it was one of the first types implemented when this
crate was originally written.

`put_auth_key` now takes an `Into<AuthKey>` as an argument (as well as
`delegated_capabilities, fixing #68) which should make it easy to insert
password-derived `AuthKey`s into HSMs.

Finally, MockHSM has been updated to support `AuthKey`s, and now handles
them the same way it would any other object.