Handle session timeouts
Sessions with the YubiHSM2 are stateful and time out after 30 seconds of
inactivity (as tracked by the HSM's internal clock). This manifests as
errors from the HSM which are difficult to distinguish from other error
conditions.

This commit tracks how long it's been since we've sent a command and
returns a `TimeoutError` in the event it's been too long (30 seconds
minus 1 second skew).