Refactor Adapter trait; add Connection type
This changes the API of the Adapter trait to be less `yubihsm-connector`
centric, simplifying it down to just:

- open: make a connection to the YubiHSM
- is_open: are we still connected to the YubiHSM
- send_message: send a message to the YubiHSM and return the response

Additionally, it splits apart `Session` into separate `Session` and
`Connection` types. The `Connection` type handles managing the
underlying `Adapter` and initializing the `SecureChannel`. The `Session`
encrypts messages using the `Connection`'s underlying channel, and also
handles errors, reinitalizing the underlying `Connection` if it becomes
unhealthy.