Factor `integration.rs` tests into separate submodules
The previous `integration.rs` was a gigantic file full of the entire
test suite. It was a bit hard to tell what was going on.

This factors the tests into individual submodules for each command.
Notable is that there is a 1:1 mapping between:

```
tests/commands/*.rs
```

and

```
src/commands/*.rs
```

That is to say, for each test submodule in `tests/commands/*.rs`, there
is a corresponding module in `src/commands` whose name maps to the
YubiHSM2 command documented here:

<https://developers.yubico.com/YubiHSM2/Commands/>