Rename yubihsm::Session to Client
This is the first step in a large number of API changes to come with the
goal of eliminating the current `Adapter` generic parameter on the
now `Client` type.

`Session` is a bad name for what is now `Client` because it handles
reconnecting and therefore can outlive an actual session with the HSM.
The next change after this one will rename the current `Connection`
type to `Session` to free up the `Connection` name for the soon-to-be
replacement for `Adapter`. The new `Session` type will be one-shot and
1:1 with YubiHSM sessions.