Rename yubihsm::session::Connection to Session
The old `Connection` type was actually the `Session` type originally. It
got split into a different type when adding session re-connection
features to what is now `yubihsm::Client` (and the former `Session`
type). Or: there and back again.

The new `Session` type has largely the same responsibilities as the
original session type, but has been somewhat consolidated with
`SecureChannel` (now part of the `session` module).

Most of all, renaming this type back to its original name frees up the
`Connection` name for what the collective branch this commit is part of
aims to accomplish: replacing the extant `yubihsm::Client` adapter
generic with a trait object, i.e. `Connection`.