session: Catch panics which occur in drop handler
The `Session` drop handler makes a best effort to send the "close
session" command to the HSM in the event the session is presumed
healthy, so as to free up available session slots (the HSM only has 16,
so failing to do this can quickly consume them until they time out).

Unfortunately, panicking in a drop handler is really bad! To avoid that,
this catches panics and logs them.