Session abort support
Support for explicitly terminating a session without shutting it down
cleanly (i.e. sending CloseSession to the HSM)

This was added to eliminate some error messages during the Reset Device
process where a device reset is requested, and the session closed, with
the client attempting to send a CloseSession command to a device that's
already halfway through wiping its own memory.

It's also retrofitted into the parts of the Secure Channel protocol
which handle cryptographic failures (i.e. MAC mismatch), which were
already doing the same logic but in an ad hoc way.