'rsa-preview' cargo feature, secp256k1 tests, doc fixups, audit test gating
Renames the RSA cargo feature to "rsa-preview" to note its instability.

Enables secp256k1 crate in testing and docs.

Overall doc improvements.

Gate audit option testing on a cargo feature, since when tests explode
it can leave the HSM in a bad state (audit log full).

The latter could probably be addressed by clearing the audit log when
the tests start. This is noted in a comment.