Make yubihsm::client::Client thread-safe
Wraps the internal session in an `Arc<Mutex<_>>`, which is held whenever
communicating with the HSM (or attempting to connect).

This eliminates the need for the Signatory signers types to each do this
themselves internally.

This makes `yubihsm::client::Client` Clone-able, and eliminates the need
to take a `&mut` reference to it whenever making any calls to it.