setup: Support skipping initial device reset
This refactors the setup process so as to support skipping an initial
device reset.

It seems certain YubiHSM2s (possibly from a bad batch) have trouble
completing the reset process. In cases like that, it would be nice to
provide an option to not reset the device.

This also makes the `reset_device_and_reconnect` method a public method
of `yubihsm::Client`, in case someone would like to reset and reconnect,
but doesn't want to use our `yubihsm::setup` provisioning.