Retry commands after session messages limits are exceeded
The `SessionErrorKind::CommandLimitExceeded` error occurs when too many
messages have been sent in a given session, exceeding the
`MAX_COMMANDS_PER_SESSION` value (2^20).

When this limit is exceeded, an error is eagerly returned from
`encrypt_command` indicating that the current session should be closed
and a new one created with a fresh session key.

Previously these errors bubbled up and were returned from the client.
This commit retries the command by initiating a new session and,
if successful, resending the command.

This approach is still idempotent as the error
`SessionErrorKind::CommandLimitExceeded` is eagerly returned before
any messages are sent to the HSM, ensuring that the HSM will not
receive the same command twice in this instance.