Vendor HTTP client from `harp` (#259)
The `harp` crate provided a minimalist HTTP client that does what the
YubiHSM connector client needs, and nothing more.

yubihsm.rs is currently the only user of `harp`, so it makes sense to
just vendor it into this crate so it doesn't need to be maintained
separately and refactoring can apply to both. (As it were, `harp` was
extracted from this crate with the goals of making it more general, but
that never worked out. C'est la vie)

Long term it would probably make sense to switch to a minimalist HTTP
client like `ureq`.