Remove `anomaly` dependency
The original goal of `anomaly` was to consolidate copy/pasted
error-handling boilerplate across the various iqlusion crates/apps into
a single crate which provided it instead.

In the time since, everything that was using `anomaly` has moved onto
newly emerged options for error handling like `eyre`.

While it would be good to eventually migrate this crate to `eyre` as
well, that's a fairly large undertaking. This is a first incremental
step towards that, as removing the `anomaly` dependency allows a more
piecemeal transition.

This commits vendors the parts of `anomaly` that this crate is using
into the `error.rs` file, with the goal of retiring `anomaly` entirely.