Remove session auto-close in `Drop` handler (#265)
In hopes of getting to the bottom of TMKMS issue #137, this removes the
`Drop` handler from `Session`, replacing it with a `Session::close`
method that consumes the session.

The purpose of this `Drop` handler was to inform the YubiHSM that the
session has been closed when dropped. This avoids the YubiHSM having to
wait for the session to time out, which is 30s.

However, sending messages over USB inside of a drop handler is a scary
amount of complexity, especially as panics inside of a drop handler can
lead to very hard to diagnose crashes. It tried to work around this
previously by using `catch_panic` and logging prior to aborting, but
even this is problematic.

We can potentially consider adding this back after we're confident the
existing issues have been resolved and things are otherwise stable.